requirements.txt
streamlit
pymongo
pillow
python-dotenv
matplotlib
pandas
